+++ src/utils/errors/authentication_error.ts
@@ -1,0 +1,14 @@
+import APIError from "./api_error";
+
+class AuthenticationError extends APIError {
+    constructor(instance: string, detail = "Ошибка аутентификации.") {
+        super(
+            "/errors/authentication-error", 
+            "Authentication error.", 
+            instance, 
+            detail
+        );
+    }
+}
+
+export default AuthenticationError;
