+++ src/utils/errors/database_error.ts
@@ -1,0 +1,14 @@
+import APIError from "./api_error";
+
+class DatabaseError extends APIError {
+    constructor(instance: string, detail = "Ошибка базы данных") {
+        super(
+            "/errors/database-error", 
+            "Database error.", 
+            instance, 
+            detail
+        );
+    }
+}
+
+export default DatabaseError;
